feat(briefing): extract weather pipeline (commit 4/11)
Workstream 6 Step 1 — extracts the weather pipeline from briefing-service.js
into a focused module + deletes one dead code path along the way.

NEW: server/lib/briefing/pipelines/weather.js
  - discoverWeather({ snapshot, snapshotId }) — pipeline contract entry,
    returns { weather_current, weather_forecast, reason }. Owns the dual-section
    writeSectionAndNotify (the only pipeline that writes two sections in one
    notify) and the errorMarker .catch wrapper.
  - fetchWeatherConditions({ snapshot }) — re-exported live path
    (Google Weather API). Preserves the import surface for
    server/api/briefing/briefing.js (manual refresh endpoints).
  - usesMetric, formatTemperature, formatWindSpeed,
    generateWeatherDriverImpact — now module-private.

DRIFT RESOLUTION: fetchWeatherForecast was dead code.
  - Zero callers across server/, client/, shared/.
  - Last functional change Dec 10 2025; only May 2026 touch was matrixLog tags.
  - LLM-based weather violated CLAUDE.md ABSOLUTE PRECISION ("Coordinates
    from Google APIs or DB, never from AI"). Same principle applies to
    weather data: deterministic API is authoritative.
  - Deleted along with the BRIEFING_WEATHER role in model-registry.js
    (no remaining caller) and the corresponding row in adapters/README.md.
  - Stale JSDoc example at model-registry.js:458 updated to BRIEFING_TRAFFIC.
  - Resolution logged to claude_memory row #294 (category=engineering-pattern).

briefing-service.js delta (3,178 → 2,907, net -271 lines):
  - import { discoverWeather, fetchWeatherConditions } from pipelines/weather.js
  - export { fetchWeatherConditions } (re-export for external surface)
  - orchestrator weatherPromise: replaced .then/.catch chain with direct
    discoverWeather() call (~7-line shrink, pipeline owns its own SSE write
    and errorMarker)
  - final-assembly: weatherResult?.current → weather_current,
    weatherResult?.forecast → weather_forecast (3 references at lines
    2560/2572/2592)
  - deleted ~270 lines of weather code (1 dead function, 4 helpers, 1 live)

Verification:
  - npm run typecheck (tsc --noEmit): clean
  - npm run lint (eslint client/src): clean
  - node --check on briefing-service.js, pipelines/weather.js, model-registry.js: pass
  - Dynamic import smoke test: weather.js exports
    { discoverWeather, fetchWeatherConditions }
  - Grep audits: zero orphan local definitions, zero literal channel strings,
    re-export chain resolves for external callers (server/api/briefing/briefing.js)
  - Both error pathways traced and preserved:
    Pathway A (thrown) → Promise.allSettled at 2493 captures → failedReasons.weather
    populated → fallback at 2572 fires
    Pathway B (graceful) → weatherResult populated with new shape →
    weatherResult.weather_current resolves at 2572

DOCS:
  - docs/review-queue/PLAN_workstream6_commit4_weather-2026-05-02.md
    (NEW: 884-line implementation plan written this session)
  - docs/review-queue/PLAN_workstream6_remaining-2026-05-02.md
    (UPDATED: §3 commit 4 now has "✅ RESOLVED 2026-05-02 (Option A)" note
    documenting the dead-code removal)

Per PLAN_workstream6_remaining-2026-05-02.md §2.1 (wider contract),
§2.5 (re-export pattern), §2.3 (sed for line-range deletes).
